refactor: move older privacy modals to their own folder
