+++ dev-client/src/components/modals/privacy/PrivacyInfoContent.tsx
